Google login component added in client, routers for google login created for server
