--- .env.example
@@ -1,20 +1,0 @@
-
-# Put lots of randomness in these
-SESSION_SECRET=ashdfjhasdlkjfhalksdjhflak
-JWT_SECRET=ashdfjhasdlkjfhalksdjhflak
-
-# Google keys - register your app and get yours here: https://console.cloud.google.com/
-GOOGLE_CLIENT_ID=665456456-asd351asda861casd6asd51asczxc.apps.googleusercontent.com
-GOOGLE_CLIENT_SECRET=Asd864asdAsdasd-
-
-# Application Port - express server listens on this port (default 3000).
-PORT=5000
-
-# Postgres variables
-PG_USER=postgres
-PG_HOST=localhost 
-PG_DB=dbName
-PG_PW=password
-PG_PORT=5432
-PG_URI=postgres://<host>:<port>/<key>
-PG_URI_LOCAL=postgres://<host>:<port>/<key>
