+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/index.tsx":"1","/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/redux/reducers/index.ts":"2","/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/App.tsx":"3","/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/components/UserLogin/index.tsx":"4"},{"size":509,"mtime":1606385654641,"results":"5","hashOfConfig":"6"},{"size":265,"mtime":1606385654641,"results":"7","hashOfConfig":"6"},{"size":284,"mtime":1606465008844,"results":"8","hashOfConfig":"6"},{"size":845,"mtime":1606476218837,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ayfheq",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/index.tsx",[],"/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/redux/reducers/index.ts",[],"/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/App.tsx",[],"/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/components/UserLogin/index.tsx",[]]
+[{"/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/App.tsx":"1","/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/components/GoogleUserLogin/index.tsx":"2"},{"size":302,"mtime":1606483528798,"results":"3","hashOfConfig":"4"},{"size":835,"mtime":1606484389990,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ayfheq",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/App.tsx",[],"/home/chiranjibi/Desktop/Integrify/PROJECTS/JoinMe/client/src/components/GoogleUserLogin/index.tsx",[]]
