+++ package.json
@@ -32,7 +32,6 @@
     "@types/jest": "^26.0.15",
     "@types/jsonwebtoken": "^8.5.0",
     "@types/lusca": "^1.6.2",
-    "@types/passport": "^1.0.4",
     "@types/pg": "^7.14.7",
     "@types/react-dom": "^17.0.0",
     "@types/react-redux": "^7.1.11",
@@ -57,7 +56,6 @@
     "knex": "^0.21.12",
     "lusca": "^1.6.1",
     "nodemon": "^2.0.6",
-    "passport": "^0.4.1",
     "passport-google-id-token": "^0.4.7",
     "pg": "^8.5.1",
     "react-google-login": "^5.1.25",
